refactor: split into library + binary crates (#23)
